lint

+++ libbeat/publisher/pipeline/controller_test.go
@@ -106,7 +106,7 @@ func TestOutputReload(t *testing.T) {
 
 				timeout := 20 * time.Second
 				return waitUntilTrue(timeout, func() bool {
-					return uint(numEventsToPublish) == publishedCount.Load()
+					return numEventsToPublish == publishedCount.Load()
 				})
 			}, &quick.Config{MaxCount: 25})
